Work in progress changes to build with hipsycl. TODO add support for atomic accessors to hipsycl

+++ Core/SExpr.cpp
@@ -10,7 +10,7 @@ SExpr::SExpr()
 SExpr::SExpr(const std::shared_ptr<std::string> &literalValue) {
   try {
     const Float_t tempVal =
-        static_cast<Float_t>(std::stod(literalValue->c_str()));
+        static_cast<Float_t>(std::stof(literalValue->c_str()));
     m_type = Type::Number;
     m_numValue = tempVal;
   } catch (const std::exception &) {
