+++ Core/Types.h
@@ -2,5 +2,5 @@
 
 namespace FunGPU {
 using Index_t = unsigned int;
-using Float_t = double;
+using Float_t = float;
 } // namespace FunGPU
